Move storage classes into dedicated module
