+++ spec/lib/tasks/data_export_spec.rb
@@ -5,7 +5,7 @@ describe "rake core:data_export", type: task do
   subject(:task) { Rake::Task["core:data_export"] }
 
   let(:paas_instance) { "paas_export_instance" }
-  let(:storage_service) { instance_double(S3StorageService) }
+  let(:storage_service) { instance_double(Storage::S3Service) }
   let(:paas_config_service) { instance_double(PaasConfigurationService) }
   let(:export_service) { instance_double(Exports::CaseLogExportService) }
 
@@ -14,7 +14,7 @@ describe "rake core:data_export", type: task do
     Rake::Task.define_task(:environment)
     task.reenable
 
-    allow(S3StorageService).to receive(:new).and_return(storage_service)
+    allow(Storage::S3Service).to receive(:new).and_return(storage_service)
     allow(PaasConfigurationService).to receive(:new).and_return(paas_config_service)
     allow(Exports::CaseLogExportService).to receive(:new).and_return(export_service)
     allow(ENV).to receive(:[])
@@ -23,7 +23,7 @@ describe "rake core:data_export", type: task do
 
   context "when exporting case logs with no parameters" do
     it "starts the XML export process" do
-      expect(S3StorageService).to receive(:new).with(paas_config_service, paas_instance)
+      expect(Storage::S3Service).to receive(:new).with(paas_config_service, paas_instance)
       expect(Exports::CaseLogExportService).to receive(:new).with(storage_service)
       expect(export_service).to receive(:export_xml_case_logs)
 
@@ -33,7 +33,7 @@ describe "rake core:data_export", type: task do
 
   context "when exporting case logs with CSV format" do
     it "starts the CSV export process" do
-      expect(S3StorageService).to receive(:new).with(paas_config_service, paas_instance)
+      expect(Storage::S3Service).to receive(:new).with(paas_config_service, paas_instance)
       expect(Exports::CaseLogExportService).to receive(:new).with(storage_service)
       expect(export_service).to receive(:export_csv_case_logs)
