+++ spec/lib/tasks/full_import_spec.rb
@@ -5,8 +5,8 @@ require "zip"
 describe "rake core:full_import", type: :task do
   subject(:task) { Rake::Task["core:full_import"] }
 
-  let(:s3_service) { instance_double(S3StorageService) }
-  let(:archive_service) { instance_double(ArchiveStorageService) }
+  let(:s3_service) { instance_double(Storage::S3Service) }
+  let(:archive_service) { instance_double(Storage::ArchiveService) }
   let(:paas_config_service) { instance_double(PaasConfigurationService) }
 
   before do
@@ -15,9 +15,9 @@ describe "rake core:full_import", type: :task do
     task.reenable
 
     allow(PaasConfigurationService).to receive(:new).and_return(paas_config_service)
-    allow(S3StorageService).to receive(:new).and_return(s3_service)
+    allow(Storage::S3Service).to receive(:new).and_return(s3_service)
     allow(s3_service).to receive(:get_file_io)
-    allow(ArchiveStorageService).to receive(:new).and_return(archive_service)
+    allow(Storage::ArchiveService).to receive(:new).and_return(archive_service)
   end
 
   context "when starting a full import" do
