+++ spec/services/imports/case_logs_import_service_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe Imports::CaseLogsImportService do
   subject(:case_log_service) { described_class.new(storage_service, logger) }
 
-  let(:storage_service) { instance_double(S3StorageService) }
+  let(:storage_service) { instance_double(Storage::S3Service) }
   let(:logger) { instance_double(ActiveSupport::Logger) }
 
   let(:real_2021_2022_form) { Form.new("config/forms/2021_2022.json", "2021_2022") }
