+++ spec/services/imports/scheme_import_service_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe Imports::SchemeImportService do
   subject(:scheme_service) { described_class.new(storage_service, logger) }
 
-  let(:storage_service) { instance_double(S3StorageService) }
+  let(:storage_service) { instance_double(Storage::S3Service) }
   let(:logger) { instance_double(ActiveSupport::Logger) }
 
   let(:fixture_directory) { "spec/fixtures/imports/mgmtgroups" }
